refactor(ticket) : 티켓아이템관련 로직 분리
