+++ .github/pull_request_template.md
@@ -1,6 +1,6 @@
 ### Related Issues
-<!-- 예시) * resolves #71 -->
-
+<!-- - close 뒤에 이슈달면 링크 걸리고 이슈 닫혀요 -->
+- close
 ### Describe your changes
 <!-- 작업내용을 적어주세요 -->
